Backporting edits from CGAP
