+++ src/encoded/static/components/item-pages/components/index.js
@@ -16,6 +16,7 @@ export { PartialList } from './PartialList';
 export { ExternalReferenceLink } from './ExternalReferenceLink';
 
 export { FilesInSetTable, FileItemRow } from './FilesInSetTable';
+export { ItemFileAttachment } from './ItemFileAttachment';
 export { ItemFooterRow } from './ItemFooterRow';
 export { Publications } from './Publications';
 export { SOPBelowHeaderRow } from './SOPBelowHeaderRow';
